Rename pkg from trie to tree
