+++ internal/api/clientapi/standard_discover.go
@@ -7,7 +7,7 @@ import (
 	"github.com/ConsenSys/fc-retrieval-gateway/internal/gateway"
 	"github.com/ConsenSys/fc-retrieval-gateway/internal/util"
 	"github.com/ConsenSys/fc-retrieval-gateway/pkg/fcrcrypto"
-	"github.com/ConsenSys/fc-retrieval-gateway/pkg/fcrmerkletrie"
+	"github.com/ConsenSys/fc-retrieval-gateway/pkg/fcrmerkletree"
 	"github.com/ConsenSys/fc-retrieval-gateway/pkg/fcrmessages"
 	"github.com/ConsenSys/fc-retrieval-gateway/pkg/logging"
 	"github.com/ant0ine/go-json-rest/rest"
@@ -36,13 +36,13 @@ func handleClientStandardCIDDiscover(w rest.ResponseWriter, request *fcrmessages
 	offers, exists := g.Offers.GetOffers(pieceCID)
 
 	roots := make([]string, 0)
-	proofs := make([]fcrmerkletrie.FCRMerkleProof, 0)
+	proofs := make([]fcrmerkletree.FCRMerkleProof, 0)
 	fundedPaymentChannel := make([]bool, 0)
 
 	for _, offer := range offers {
-		trie := offer.GetMerkleTrie()
-		roots = append(roots, trie.GetMerkleRoot())
-		proof, err := trie.GenerateMerkleProof(pieceCID)
+		tree := offer.GetMerkleTrie()
+		roots = append(roots, tree.GetMerkleRoot())
+		proof, err := tree.GenerateMerkleProof(pieceCID)
 		if err != nil {
 			s := "Internal error: Error generating proof."
 			logging.Error(s + err.Error())
