+++ pkg/fcrmerkletree/tree.go
@@ -1,4 +1,4 @@
-package fcrmerkletrie
+package fcrmerkletree
 
 import (
 	"github.com/cbergoon/merkletree"
@@ -6,26 +6,26 @@ import (
 
 // FCRMerkleTrie is used to store
 type FCRMerkleTrie struct {
-	trie *merkletree.MerkleTree
+	tree *merkletree.MerkleTree
 }
 
-// CreateMerkleTrie creates a merkle trie from a list of cids
+// CreateMerkleTrie creates a merkle tree from a list of cids
 func CreateMerkleTrie(contents []merkletree.Content) (*FCRMerkleTrie, error) {
-	trie, err := merkletree.NewTree(contents)
+	tree, err := merkletree.NewTree(contents)
 	if err != nil {
 		return nil, err
 	}
-	return &FCRMerkleTrie{trie: trie}, nil
+	return &FCRMerkleTrie{tree: tree}, nil
 }
 
-// GetMerkleRoot returns the merkle root of the trie
+// GetMerkleRoot returns the merkle root of the tree
 func (mt *FCRMerkleTrie) GetMerkleRoot() string {
-	return string(mt.trie.MerkleRoot())
+	return string(mt.tree.MerkleRoot())
 }
 
 // GenerateMerkleProof gets the merkle proof for a given cid
 func (mt *FCRMerkleTrie) GenerateMerkleProof(content merkletree.Content) (*FCRMerkleProof, error) {
-	path, index, err := mt.trie.GetMerklePath(content)
+	path, index, err := mt.tree.GetMerklePath(content)
 	if err != nil {
 		return nil, err
 	}
